Separate ui/server utils
